Refactor: Move NIM-specific model format extraction logic from kserve package to frontend wrapper to maintain clean package separation
