+++ packages/kserve/src/modelFormat.ts
@@ -1,9 +1,4 @@
 import type { SupportedModelFormats } from '@odh-dashboard/internal/k8sTypes';
-// eslint-disable-next-line @odh-dashboard/no-restricted-imports
-import {
-  isNIMOperatorManaged,
-  getModelNameFromNIMInferenceService,
-} from '@odh-dashboard/internal/pages/modelServing/screens/global/nimOperatorUtils';
 import type { KServeDeployment } from './deployments';
 
 export const extractKServeModelFormat = (
@@ -15,17 +10,6 @@ export const extractKServeModelFormat = (
       name: modelFormat.name,
       version: modelFormat.version,
     };
-  }
-
-  // Check if this is a NIM Operator-managed deployment
-  if (isNIMOperatorManaged(deployment.model)) {
-    const modelName = getModelNameFromNIMInferenceService(deployment.model);
-    if (modelName) {
-      return {
-        name: modelName,
-        version: undefined,
-      };
-    }
   }
 
   return null;
